Solution to Exercise 2
I have changed the way the labels are constructed for different alpha values in the legend. In the hint for the exercise 2 you mentioned string concatenation as well as str() function for constructing labels for different alpha values, however in the solution you used f string. So, may be it would be better to show first how string concatenation is sed and then propose f string as an alternative version.